requirements.txt
Django==4.2.3
django-environ==0.10.0
django-crispy-forms==2.0
crispy_bootstrap5==2022.1
django-taggit==4.0.0